for_app2/requirements.txt
numpy==1.19.1
Keras==2.4.3
streamlit==0.69.2
pandas==1.1.2
Pillow==8.0.1
Tensorflow-cpu == 2.3.1
h5py == 2.10.0